requirements.txt
grpcio==1.27.2
grpcio-tools==1.27.2
pkg-resources==0.0.0
protobuf==3.11.3
six==1.14.0
